upgraded simple-history and fixed department models.
