+++ coldfront/core/department/models.py
@@ -9,3 +9,15 @@ class Department(models.Model):
 
     def __str__(self):
         return f'{self.name} ({self.acronym})'
+
+class UserDepartment(models.Model):
+    user_profile = models.ForeignKey('user.UserProfile', on_delete=models.CASCADE)
+    department = models.ForeignKey(Department, on_delete=models.CASCADE)
+    is_authoritative = models.BooleanField(default=False)
+    history = HistoricalRecords()
+
+    def __str__(self):
+        return f'{self.user_profile.user.username}-{self.department.acronym}'
+
+    class Meta:
+        unique_together = ('user_profile', 'department')
