+++ requirements.txt
@@ -18,7 +18,7 @@ django-phonenumber-field==5.1.0
 django-picklefield==2.0
 django-q==1.0.1
 django-settings-export==1.2.1
-django-simple-history==2.12.0
+django-simple-history==3.2.0
 django-sslserver==0.20
 django-su==1.0.0
 djangorestframework==3.11.1
